+++ l10n_br_account/demo/base_demo.xml
@@ -6,12 +6,12 @@
 		<record id="l10n_br_base.res_partner_akretion" model="res.partner">
 		    	<field name="partner_fiscal_type_id" ref="l10n_br_account.partner_fiscal_type_1" />
 		</record>
-		
+
 		<!-- CHINEN BOX -->
 		<record id="l10n_br_base.res_partner_chinenbox0" model="res.partner">
 		    	<field name="partner_fiscal_type_id" ref="l10n_br_account.partner_fiscal_type_1" />
 		</record>
-		
+
 		<!-- CYBER CAFE -->
 		<record id="l10n_br_base.res_partner_cybercafe0" model="res.partner">
 		    	<field name="partner_fiscal_type_id" ref="l10n_br_account.partner_fiscal_type_1" />
@@ -46,7 +46,7 @@
 		<record id="base.res_partner_5" model="res.partner">
 		    	<field name="partner_fiscal_type_id" ref="l10n_br_account.partner_fiscal_type_1" />
 		</record>
-	
+
 		<!-- Bank Wealthy and sons -->
 		<record id="base.res_partner_2" model="res.partner">
 		    	<field name="partner_fiscal_type_id" ref="l10n_br_account.partner_fiscal_type_1" />
