+++ l10n_br_account/l10n_br_account.py
@@ -17,62 +17,78 @@
 #along with this program.  If not, see <http://www.gnu.org/licenses/>.        #
 ###############################################################################
 
-import re
-
-from l10n_br_base.tools import fiscal
 from openerp.osv import orm, fields
+from openerp import netsvc
+
+TYPE = [
+    ('input', 'Entrada'),
+    ('output', u'Saída'),
+]
+
+PRODUCT_FISCAL_TYPE = [
+    ('service', u'Serviço'),
+]
+
+PRODUCT_FISCAL_TYPE_DEFAULT = PRODUCT_FISCAL_TYPE[0][0]
 
 
-class l10n_br_account_cfop(orm.Model):
-    """ CFOP - Código Fiscal de Operações e Prestações """
-    _name = 'l10n_br_account.cfop'
-    _description = 'CFOP'
+class L10n_brAccountFiscalCategory(orm.Model):
+    _name = 'l10n_br_account.fiscal.category'
+    _description = 'Categoria Fiscail'
     _columns = {
-        'code': fields.char(u'Código', size=4, required=True),
-        'name': fields.char('Nome', size=256, required=True),
-        'small_name': fields.char('Nome Reduzido', size=32, required=True),
-        'description': fields.text(u'Descrição'),
-        'type': fields.selection(
-            [('input', 'Entrada'), ('output', u'Saída')], 'Tipo',
-            required=True),
-        'parent_id': fields.many2one('l10n_br_account.cfop', 'CFOP Pai'),
-        'child_ids': fields.one2many(
-            'l10n_br_account.cfop', 'parent_id', 'CFOP Filhos'),
-        'internal_type': fields.selection(
-            [('view', u'Visualização'), ('normal', 'Normal')],
-            'Tipo Interno', required=True),
+        'code': fields.char(u'Código', size=254, required=True),
+        'name': fields.char(u'Descrição', size=254),
+        'type': fields.selection(TYPE, 'Tipo'),
+        'fiscal_type': fields.selection(PRODUCT_FISCAL_TYPE, 'Tipo Fiscal'),
+        'property_journal': fields.property(
+            'account.journal', type='many2one', relation='account.journal',
+            string=u"Diário Contábil", method=True, view_load=True,
+            help=u"Diário utilizado para esta categoria de operação fiscal"),
+        'journal_type': fields.selection(
+            [('sale', 'Venda'), ('sale_refund', u'Devolução de Venda'),
+            ('purchase', 'Compras'),
+            ('purchase_refund', u'Devolução de Compras')], u'Tipo do Diário',
+            size=32, required=True),
+        'refund_fiscal_category_id': fields.many2one(
+            'l10n_br_account.fiscal.category', u'Categoria Fiscal de Devolução',
+            domain="""[('type', '!=', type), ('fiscal_type', '=', fiscal_type),
+                ('journal_type', 'like', journal_type),
+                ('state', '=', 'approved')]"""),
+        'reverse_fiscal_category_id': fields.many2one(
+            'l10n_br_account.fiscal.category', u'Categoria Fiscal Inversa',
+            domain="""[('type', '!=', type), ('fiscal_type', '=', fiscal_type),
+                ('state', '=', 'approved')]"""),
+        'fiscal_position_ids': fields.one2many('account.fiscal.position',
+            'fiscal_category_id', u'Posições Fiscais'),
+        'note': fields.text(u'Observações'),
+        'state': fields.selection([('draft', u'Rascunho'),
+            ('review', u'Revisão'), ('approved', u'Aprovada'),
+            ('unapproved', u'Não Aprovada')], 'Status', readonly=True,
+            track_visibility='onchange', select=True),
     }
     _defaults = {
-        'internal_type': 'normal',
+        'state': 'draft',
+        'type': 'output',
+        'fiscal_type': PRODUCT_FISCAL_TYPE_DEFAULT,
+        'journal_type': 'sale',
     }
     _sql_constraints = [
-        ('l10n_br_account_cfop_code_uniq', 'unique (code)',
-         u'Já existe um CFOP com esse código !')
+        ('l10n_br_account_fiscal_category_code_uniq', 'unique (code)',
+         u'Já existe uma categoria fiscal com esse código !')
     ]
 
-    def name_search(self, cr, user, name, args=None, operator='ilike',
-                    context=None, limit=80):
-        if not args:
-            args = []
-        if context is None:
-            context = {}
-        ids = self.search(cr, user, ['|', ('name', operator, name),
-                                     ('code', operator, name)] + args,
-                          limit=limit, context=context)
-        return self.name_get(cr, user, ids, context)
-
-    def name_get(self, cr, uid, ids, context=None):
-        if not len(ids):
-            return []
-        if isinstance(ids, (int, long)):
-            ids = [ids]
-        reads = self.read(cr, uid, ids, ['name', 'code'], context,
-                          load='_classic_write')
-        return [(x['id'], (x['code'] and x['code'] or '') +
-                 (x['name'] and ' - ' + x['name'] or '')) for x in reads]
+    def action_unapproved_draft(self, cr, uid, ids, *args):
+        self.write(cr, uid, ids, {'state': 'draft'})
+        wf_service = netsvc.LocalService("workflow")
+        for fc_id in ids:
+            wf_service.trg_delete(
+                uid, 'l10n_br_account.fiscal.category', fc_id, cr)
+            wf_service.trg_create(
+                uid, 'l10n_br_account.fiscal.category', fc_id, cr)
+        return True
 
 
-class l10n_br_account_service_type(orm.Model):
+class L10n_brAccountServiceType(orm.Model):
     _name = 'l10n_br_account.service.type'
     _description = u'Cadastro de Operações Fiscais de Serviço'
     _columns = {
@@ -104,7 +120,7 @@ class l10n_br_account_service_type(orm.Model):
         return res
 
 
-class l10n_br_account_fiscal_document(orm.Model):
+class L10n_brAccountFiscalDocument(orm.Model):
     _name = 'l10n_br_account.fiscal.document'
     _description = 'Tipo de Documento Fiscal'
     _columns = {
@@ -114,64 +130,25 @@ class l10n_br_account_fiscal_document(orm.Model):
     }
 
 
-class l10n_br_account_fiscal_category(orm.Model):
-    _name = 'l10n_br_account.fiscal.category'
-    _description = 'Categoria Fiscail'
-    _columns = {
-        'code': fields.char(u'Código', size=254, required=True),
-        'name': fields.char(u'Descrição', size=254),
-        'type': fields.selection(
-            [('input', 'Entrada'), ('output', u'Saída')], 'Tipo'),
-        'property_journal': fields.property(
-            'account.journal', type='many2one', relation='account.journal',
-            string=u"Diário Contábil", method=True, view_load=True,
-            help=u"Diário utilizado para esta categoria de operação fiscal"),
-        'fiscal_type': fields.selection(
-            [('product', 'Produto'), ('service', u'Serviço')],
-            'Tipo Fiscal', required=True),
-        'journal_type': fields.selection(
-            [('sale', 'Venda'), ('sale_refund', u'Devolução de Venda'),
-            ('purchase', 'Compras'),
-            ('purchase_refund', u'Devolução de Compras')], u'Tipo do Diário',
-            size=32, required=True),
-        'refund_fiscal_category_id': fields.many2one(
-            'l10n_br_account.fiscal.category',
-            u'Categoria Fiscal de Devolução',
-            domain="[('type', '!=', type), ('fiscal_type', '=', fiscal_type), "
-            "('journal_type', 'like', journal_type)]"),
-        'fiscal_position_ids': fields.one2many('account.fiscal.position',
-                                               'fiscal_category_id',
-                                               u'Posições Fiscais'),
-        'note': fields.text(u'Observações')
-    }
-    _defaults = {
-        'type': 'output',
-        'fiscal_type': 'product',
-        'journal_type': 'sale'
-    }
-
-
-class l10n_br_account_document_serie(orm.Model):
+class L10n_brAccountDocumentSerie(orm.Model):
     _name = 'l10n_br_account.document.serie'
     _description = 'Serie de documentos fiscais'
     _columns = {
         'code': fields.char(u'Código', size=3, required=True),
         'name': fields.char(u'Descrição', size=64, required=True),
+        'fiscal_type': fields.selection(PRODUCT_FISCAL_TYPE, 'Tipo Fiscal'),
         'fiscal_document_id': fields.many2one(
             'l10n_br_account.fiscal.document',
             'Documento Fiscal', required=True),
         'company_id': fields.many2one(
             'res.company', 'Empresa', required=True),
         'active': fields.boolean('Ativo'),
-        'fiscal_type': fields.selection(
-            [('product', 'Produto'), ('service', u'Serviço')],
-            'Tipo Fiscal', required=True),
         'internal_sequence_id': fields.many2one(
             'ir.sequence', u'Sequência Interna')
     }
     _defaults = {
         'active': True,
-        'fiscal_type': 'product'
+        'fiscal_type': PRODUCT_FISCAL_TYPE_DEFAULT,
     }
 
     def create_sequence(self, cr, uid, vals, context=None):
@@ -193,7 +170,7 @@ class l10n_br_account_document_serie(orm.Model):
         not vals['internal_sequence_id']:
             vals.update({'internal_sequence_id': self.create_sequence(
                 cr, uid, vals, context)})
-        result = super(l10n_br_account_document_serie, self).create(
+        result = super(L10n_brAccountDocumentSerie, self).create(
             cr, uid, vals, context)
         if result:
             company = self.pool.get('res.company').browse(
@@ -316,7 +293,7 @@ class l10n_br_account_invoice_invalid_number(orm.Model):
         return True
 
 
-class l10n_br_account_partner_fiscal_type(orm.Model):
+class L10n_brAccountPartnerFiscalType(orm.Model):
     _name = 'l10n_br_account.partner.fiscal.type'
     _description = 'Tipo Fiscal de Parceiros'
     _columns = {
@@ -359,7 +336,7 @@ class l10n_br_account_cnae(orm.Model):
         return res
 
 
-class l10n_br_tax_definition_template(orm.Model):
+class L10n_brTaxDefinitionTemplate(orm.Model):
     _name = 'l10n_br_tax.definition.template'
     _columns = {
         'tax_id': fields.many2one(
@@ -377,7 +354,7 @@ class l10n_br_tax_definition_template(orm.Model):
         return {'value': {'tax_domain': tax_domain}}
 
 
-class l10n_br_tax_definition(orm.Model):
+class L10n_brTaxDefinition(orm.Model):
     _name = 'l10n_br_tax.definition'
     _columns = {
         'tax_id': fields.many2one('account.tax', 'Imposto', required=True),
@@ -396,170 +373,3 @@ class l10n_br_tax_definition(orm.Model):
             tax_domain = self.pool.get('account.tax').read(
                 cr, uid, tax_id, ['domain'], context=context)['domain']
         return {'value': {'tax_domain': tax_domain}}
-
-
-class l10n_br_account_document_related(orm.Model):
-    _name = 'l10n_br_account.document.related'
-    _columns = {
-        'invoice_id': fields.many2one(
-            'account.invoice', 'Documento Fiscal',
-            ondelete='cascade', select=True),
-        'invoice_related_id': fields.many2one(
-            'account.invoice', 'Documento Fiscal',
-            ondelete='cascade', select=True),
-        'document_type': fields.selection(
-            [('nf', 'NF'), ('nfe', 'NF-e'), ('cte', 'CT-e'),
-                ('nfrural', 'NF Produtor'), ('cf', 'Cupom Fiscal')],
-            'Tipo Documento', required=True),
-        'access_key': fields.char('Chave de Acesso', size=44),
-        'serie': fields.char(u'Série', size=12),
-        'internal_number': fields.char(u'Número', size=32),
-        'state_id': fields.many2one(
-            'res.country.state', 'Estado',
-            domain="[('country_id.code', '=', 'BR')]"),
-        'cnpj_cpf': fields.char('CNPJ/CPF', size=18),
-        'cpfcnpj_type': fields.selection(
-            [('cpf', 'CPF'), ('cnpj', 'CNPJ')], 'Tipo Doc.'),
-        'inscr_est': fields.char('Inscr. Estadual/RG', size=16),
-        'date': fields.date('Data'),
-        'fiscal_document_id': fields.many2one(
-            'l10n_br_account.fiscal.document', 'Documento'),
-    }
-    _defaults = {
-        'cpfcnpj_type': 'cnpj',
-    }
-
-    def _check_cnpj_cpf(self, cr, uid, ids):
-
-        for inv_related in self.browse(cr, uid, ids):
-            if not inv_related.cnpj_cpf:
-                continue
-
-            if inv_related.cpfcnpj_type == 'cnpj':
-                if not fiscal.validate_cnpj(inv_related.cnpj_cpf):
-                    return False
-            elif not fiscal.validate_cpf(inv_related.cnpj_cpf):
-                    return False
-
-        return True
-
-    def _check_ie(self, cr, uid, ids):
-        """Checks if company register number in field insc_est is valid,
-        this method call others methods because this validation is State wise
-
-        :Return: True or False.
-
-        :Parameters:
-            - 'cr': Database cursor.
-            - 'uid': Current user’s ID for security checks.
-            - 'ids': List of partner objects IDs.
-        """
-        for inv_related in self.browse(cr, uid, ids):
-            if not inv_related.inscr_est \
-            or inv_related.inscr_est == 'ISENTO':
-                continue
-
-            uf = inv_related.state_id and \
-            inv_related.state_id.code.lower() or ''
-
-            try:
-                mod = __import__(
-                'l10n_br_base.tools.fiscal', globals(), locals(), 'fiscal')
-
-                validate = getattr(mod, 'validate_ie_%s' % uf)
-                if not validate(inv_related.inscr_est):
-                    return False
-            except AttributeError:
-                if not fiscal.validate_ie_param(uf, inv_related.inscr_est):
-                    return False
-
-        return True
-
-    _constraints = [
-        (_check_cnpj_cpf, u'CNPJ/CPF do documento relacionado é invalido!',
-            ['cnpj_cpf']),
-        (_check_ie, u'Inscrição Estadual do documento fiscal inválida!',
-            ['inscr_est']),
-    ]
-
-    def onchange_invoice_related_id(self, cr, uid, ids,
-                                    invoice_related_id=False, context=None):
-        result = {'value': {}}
-
-        if not invoice_related_id:
-            return result
-
-        inv_related = self.pool.get('account.invoice').browse(
-            cr, uid, invoice_related_id)
-
-        if not inv_related.fiscal_document_id:
-            return result
-
-        if inv_related.fiscal_document_id.code == '01':
-            result['value']['document_type'] = 'nf'
-        elif inv_related.fiscal_document_id.code == '04':
-            result['value']['document_type'] = 'nfrural'
-        elif inv_related.fiscal_document_id.code == '55':
-            result['value']['document_type'] = 'nfe'
-        elif inv_related.fiscal_document_id.code == '57':
-            result['value']['document_type'] = 'cte'
-        elif inv_related.fiscal_document_id.code in ('2B', '2C', '2D'):
-            result['value']['document_type'] = 'cf'
-        else:
-            result['value']['document_type'] = False
-
-        if inv_related.fiscal_document_id.code in ('55', '57'):
-            result['value']['access_key'] = inv_related.nfe_access_key
-            result['value']['serie'] = False
-            result['value']['serie'] = False
-            result['value']['internal_number'] = False
-            result['value']['state_id'] = False
-            result['value']['cnpj_cpf'] = False
-            result['value']['cpfcnpj_type'] = False
-            result['value']['date'] = False
-            result['value']['fiscal_document_id'] = False
-            result['value']['inscr_est'] = False
-
-        if inv_related.fiscal_document_id.code in ('01', '04'):
-            result['value']['access_key'] = False
-            if inv_related.issuer == '0':
-                result['value']['serie'] = inv_related.document_serie_id and \
-                inv_related.document_serie_id.code or False
-            else:
-                result['value']['serie'] = inv_related.vendor_serie
-
-            result['value']['internal_number'] = inv_related.internal_number
-            result['value']['state_id'] = inv_related.partner_id and \
-            inv_related.partner_id.state_id and \
-            inv_related.partner_id.state_id.id or False
-            result['value']['cnpj_cpf'] = inv_related.partner_id and \
-            inv_related.partner_id.cnpj_cpf or False
-
-            if inv_related.partner_id.is_company:
-                result['value']['cpfcnpj_type'] = 'cnpj'
-            else:
-                result['value']['cpfcnpj_type'] = 'cpf'
-
-            result['value']['date'] = inv_related.date_invoice
-            result['value']['fiscal_document_id'] = inv_related.fiscal_document_id and \
-            inv_related.fiscal_document_id.id or False
-
-        if inv_related.fiscal_document_id.code == '04':
-            result['value']['inscr_est'] = inv_related.partner_id and \
-            inv_related.partner_id.inscr_est or False
-
-        return result
-
-    def onchange_mask_cnpj_cpf(self, cr, uid, ids, cpfcnpj_type, cnpj_cpf,
-                            context=None):
-        result = {'value': {}}
-        if cnpj_cpf:
-            val = re.sub('[^0-9]', '', cnpj_cpf)
-            if cpfcnpj_type == 'cnpj' and len(val) == 14:
-                cnpj_cpf = "%s.%s.%s/%s-%s"\
-                % (val[0:2], val[2:5], val[5:8], val[8:12], val[12:14])
-            elif cpfcnpj_type == 'cpf' and len(val) == 11:
-                cnpj_cpf = "%s.%s.%s-%s"\
-                % (val[0:3], val[3:6], val[6:9], val[9:11])
-            result['value'].update({'cnpj_cpf': cnpj_cpf})
-        return result
